feat: submodules created and harpoon
