+++ .gitignore
@@ -1,4 +1,5 @@
 tags
+stash.diff
 test.sh
 .luarc.json
 nvim
