--- lua/kickstart/plugins/configs/init.lua
@@ -1,1 +1,0 @@
-print('kickstart.plugins.config.init loaded')
